+++ nix/pkgs/hpio-hlint.nix
@@ -6,7 +6,7 @@
 }:
 mkDerivation {
   pname = "hpio";
-  version = "0.9.0.5";
+  version = "0.9.0.6";
   src = ../../.;
   configureFlags = [ "-ftest-hlint" ];
   isLibrary = true;