+++ nix/pkgs/hpio.nix
@@ -5,7 +5,7 @@
 }:
 mkDerivation {
   pname = "hpio";
-  version = "0.9.0.5";
+  version = "0.9.0.6";
   src = ../../.;
   isLibrary = true;
   isExecutable = true;